requirements.txt
asyncio
bleak==0.20.0
boto3==1.28.0
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4